// src/generated/ipc/UpdateRuntimeInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DistributionMethod } from "./DistributionMethod";
import type { UpdateChannel } from "./UpdateChannel";
import type { UpdateInstallStrategy } from "./UpdateInstallStrategy";

export type UpdateRuntimeInfo = {
  currentVersion: string;
  channel: UpdateChannel;
  channelMutable: boolean;
  distribution: DistributionMethod;
  installStrategy: UpdateInstallStrategy;
  checkConfigured: boolean;
  installationSupported: boolean;
};